reading current timestamp as string from DB
