+++ go/inst/candidate_database_instance.go
@@ -34,7 +34,6 @@ func NewCandidateDatabaseInstance(instanceKey *InstanceKey, promotionRule Candid
 		Hostname:      instanceKey.Hostname,
 		Port:          instanceKey.Port,
 		PromotionRule: promotionRule,
-		LastSuggested: time.Now().UTC(),
 	}
 }
