+++ go/inst/candidate_database_instance_dao.go
@@ -17,10 +17,64 @@
 package inst
 
 import (
+	"fmt"
+
+	"github.com/openark/golib/log"
 	"github.com/openark/golib/sqlutils"
 
+	"github.com/github/orchestrator/go/config"
 	"github.com/github/orchestrator/go/db"
 )
+
+// RegisterCandidateInstance markes a given instance as suggested for successoring a master in the event of failover.
+func RegisterCandidateInstance(candidate *CandidateDatabaseInstance) error {
+	args := sqlutils.Args(candidate.Hostname, candidate.Port, string(candidate.PromotionRule))
+	if candidate.LastSuggested.IsZero() {
+		timeNow, err := db.ReadTimeNow()
+		if err != nil {
+			return err
+		}
+		args = append(args, timeNow)
+	} else {
+		args = append(args, candidate.LastSuggested)
+	}
+	query := fmt.Sprintf(`
+			insert into candidate_database_instance (
+					hostname,
+					port,
+					promotion_rule,
+					last_suggested
+				) values (?, ?, ?, ?)
+				on duplicate key update
+					hostname=values(hostname),
+					port=values(port),
+					last_suggested=now(),
+					promotion_rule=values(promotion_rule)
+			`)
+	writeFunc := func() error {
+		_, err := db.ExecOrchestrator(query, args...)
+		AuditOperation("register-candidate", candidate.Key(), string(candidate.PromotionRule))
+		return log.Errore(err)
+	}
+	return ExecDBWriteFunc(writeFunc)
+}
+
+// ExpireCandidateInstances removes stale master candidate suggestions.
+func ExpireCandidateInstances() error {
+	writeFunc := func() error {
+		_, err := db.ExecOrchestrator(`
+        	delete from candidate_database_instance
+				where last_suggested < NOW() - INTERVAL ? MINUTE
+				`, config.Config.CandidateInstanceExpireMinutes,
+		)
+		if err != nil {
+			return log.Errore(err)
+		}
+
+		return nil
+	}
+	return ExecDBWriteFunc(writeFunc)
+}
 
 // BulkReadCandidateDatabaseInstance returns a slice of
 // CandidateDatabaseInstance converted to JSON.
